Update edgelist.csv
According to the data example in your book, the data can have reciprocal relationships, however the same employee cannot give the same person a second rating. 

With that said, the original data had two entries with the same person as the source and the same person in the target. 
Miles,Ian,2 
Miles,Ian,3 

There should only be one entry that the follow the same person in the source and target.

How to reproduce issue: 

Chapter 7, Section 7.13 

When rebinding the edges, after updating a new edge. We get two lines that update. This is the case, because d3 looks at the key and say well we have a duplicate key, let's insert a new edge to the view.